Move getCounts thunk action to Submission/actions
- getCount is only used by Submission components
